+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atlaskit/editor-mobile-bridge
 
+## 30.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
+## 30.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
+## 30.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 30.0.1
 
 ### Patch Changes